Day 10 solutions in Matlab (@lemonad)
Made a terrible mistake today! Had a bug in part one and
didn't see that my result was not the same as the example.
However, my result was still correct for part one. Something
that resulted in a loooong wild goose chase in part two. I
must have re-read the text at least 20 times, all the time
assuming I had a correct implementation of part one :(